+++ store/core/src/main/java/org/apache/carbondata/store/impl/rpc/RegistryService.java
@@ -15,40 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.carbondata.store.impl.distributed.rpc.model;
+package org.apache.carbondata.store.impl.rpc;
 
-import java.io.DataInput;
-import java.io.DataOutput;
 import java.io.IOException;
-import java.io.Serializable;
 
 import org.apache.carbondata.common.annotations.InterfaceAudience;
+import org.apache.carbondata.store.impl.rpc.model.RegisterWorkerRequest;
+import org.apache.carbondata.store.impl.rpc.model.RegisterWorkerResponse;
 
-import org.apache.hadoop.io.Writable;
+import org.apache.hadoop.ipc.VersionedProtocol;
 
 @InterfaceAudience.Internal
-public class RegisterWorkerResponse implements Serializable, Writable {
-
-  private String workerId;
-
-  public RegisterWorkerResponse() {
-  }
-
-  public RegisterWorkerResponse(String workerId) {
-    this.workerId = workerId;
-  }
-
-  public String getWorkerId() {
-    return workerId;
-  }
-
-  @Override
-  public void write(DataOutput out) throws IOException {
-    out.writeUTF(workerId);
-  }
-
-  @Override
-  public void readFields(DataInput in) throws IOException {
-    workerId = in.readUTF();
-  }
+public interface RegistryService extends VersionedProtocol {
+  long versionID = 1L;
+  RegisterWorkerResponse registerWorker(RegisterWorkerRequest request) throws IOException;
 }
