+++ store/core/src/main/java/org/apache/carbondata/store/impl/rpc/model/BaseResponse.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.carbondata.store.impl.distributed.rpc.model;
+package org.apache.carbondata.store.impl.rpc.model;
 
 import java.io.DataInput;
 import java.io.DataOutput;
@@ -23,51 +23,47 @@ import java.io.IOException;
 import java.io.Serializable;
 
 import org.apache.carbondata.common.annotations.InterfaceAudience;
-import org.apache.carbondata.core.util.ObjectSerializationUtil;
 
 import org.apache.hadoop.io.Writable;
-import org.apache.hadoop.io.WritableUtils;
 
 @InterfaceAudience.Internal
-public class QueryResponse extends BaseResponse implements Serializable, Writable {
-  private int queryId;
-  private Object[][] rows;
+public class BaseResponse implements Serializable, Writable {
+  private int status;
+  private String message;
 
-  public QueryResponse() {
-    super();
+  public BaseResponse() {
   }
 
-  public QueryResponse(int queryId, int status, String message, Object[][] rows) {
-    super(status, message);
-    this.queryId = queryId;
-    this.rows = rows;
+  public BaseResponse(int status, String message) {
+    this.status = status;
+    this.message = message;
   }
 
-  public int getQueryId() {
-    return queryId;
+  public int getStatus() {
+    return status;
   }
 
+  public void setStatus(int status) {
+    this.status = status;
+  }
 
-  public Object[][] getRows() {
-    return rows;
+  public String getMessage() {
+    return message;
+  }
+
+  public void setMessage(String message) {
+    this.message = message;
   }
 
   @Override
   public void write(DataOutput out) throws IOException {
-    super.write(out);
-    out.writeInt(queryId);
-    WritableUtils.writeCompressedByteArray(out, ObjectSerializationUtil.serialize(rows));
+    out.writeInt(status);
+    out.writeUTF(message);
   }
 
   @Override
   public void readFields(DataInput in) throws IOException {
-    super.readFields(in);
-    queryId = in.readInt();
-    try {
-      rows = (Object[][])ObjectSerializationUtil.deserialize(
-          WritableUtils.readCompressedByteArray(in));
-    } catch (ClassNotFoundException e) {
-      throw new IOException(e);
-    }
+    status = in.readInt();
+    message = in.readUTF();
   }
 }
