+++ store/core/src/main/java/org/apache/carbondata/store/impl/worker/Worker.java
@@ -15,25 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.carbondata.store.impl.distributed;
+package org.apache.carbondata.store.impl.worker;
 
 import java.io.IOException;
 import java.net.BindException;
 
+import org.apache.carbondata.common.annotations.InterfaceAudience;
 import org.apache.carbondata.common.logging.LogService;
 import org.apache.carbondata.common.logging.LogServiceFactory;
 import org.apache.carbondata.store.api.conf.StoreConf;
-import org.apache.carbondata.store.impl.distributed.rpc.RegistryService;
-import org.apache.carbondata.store.impl.distributed.rpc.ServiceFactory;
-import org.apache.carbondata.store.impl.distributed.rpc.StoreService;
-import org.apache.carbondata.store.impl.distributed.rpc.impl.StoreServiceImpl;
-import org.apache.carbondata.store.impl.distributed.rpc.model.RegisterWorkerRequest;
-import org.apache.carbondata.store.impl.distributed.rpc.model.RegisterWorkerResponse;
+import org.apache.carbondata.store.impl.StoreServiceImpl;
+import org.apache.carbondata.store.impl.rpc.RegistryService;
+import org.apache.carbondata.store.impl.rpc.ServiceFactory;
+import org.apache.carbondata.store.impl.rpc.StoreService;
+import org.apache.carbondata.store.impl.rpc.model.RegisterWorkerRequest;
+import org.apache.carbondata.store.impl.rpc.model.RegisterWorkerResponse;
 import org.apache.carbondata.store.util.StoreUtil;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.ipc.RPC;
 
+@InterfaceAudience.Internal
 public class Worker {
 
   private static LogService LOGGER = LogServiceFactory.getLogService(Worker.class.getName());
