+++ store/horizon/src/test/java/org/apache/carbondata/horizon/HorizonTest.java
@@ -33,7 +33,7 @@ import org.apache.carbondata.horizon.rest.model.view.SelectRequest;
 import org.apache.carbondata.horizon.rest.model.view.SelectResponse;
 import org.apache.carbondata.store.api.conf.StoreConf;
 import org.apache.carbondata.store.api.exception.StoreException;
-import org.apache.carbondata.store.impl.distributed.Worker;
+import org.apache.carbondata.store.impl.worker.Worker;
 import org.apache.carbondata.store.util.StoreUtil;
 
 import org.junit.AfterClass;
